tomcat

+++ sql/department.sql
@@ -1,5 +1,5 @@
 CREATE TABLE department (
-    department_id INT AUTO_INCREMENT PRIMARY KEY,
+    department_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     department_name VARCHAR(100) NOT NULL,
     department_code CHAR(2) NOT NULL UNIQUE
 );
